Cleanup: move common wrapper component types into a shared file
